+++ elf-reader/src/main/java/pl/marcinchwedczuk/elfviewer/elfreader/elf32/sections/Elf32NotesSection.java
@@ -57,7 +57,7 @@ public class Elf32NotesSection extends Elf32Section {
                         type));
             }
 
-            curr = sf.currentPositionInFile().longValue();
+            curr = sf.currentPositionInFile().intValue();
         }
 
         return notes;
